Handle special overrides for extraLabels and extraAnnotations
In normal cases, ClusterInstance input data from the ClusterRequest overrides
the defaults from the configmap. However, for the fields extraLabels and
extraAnnotations, the default values from the configmap should take precedence
if the same labels/annotations exist in both.

Add a function to handle these special overrides. If extraLabels or extraAnnotations
are present in both the ClusterRequest and the configmap, the values from
ClusterRequest will be overriden with the defaults.

The utility merge function is invoked after to perform the normal merging behavior,
where ClusterRequest overrides defaults if both exist.